Updates data_aquisition.py to record data for the lstm

+++ data_storage/full_timewindow/data_aquisition.py
@@ -7,7 +7,6 @@ import time
 from larcc_classes.data_storage.DataForLearning import DataForLearning
 import numpy as np
 import rospy
-from lib.src.ArmGripperComm import ArmGripperComm
 
 
 def add_to_vector(data, vector, first_timestamp, dic_offset):
@@ -29,9 +28,10 @@ def add_to_vector(data, vector, first_timestamp, dic_offset):
                          data.wrench_force_torque.force.z - dic_offset["fz"],
                          data.wrench_force_torque.torque.x - dic_offset["mx"],
                          data.wrench_force_torque.torque.y - dic_offset["my"],
-                         data.wrench_force_torque.torque.z - - dic_offset["mz"]])
+                         data.wrench_force_torque.torque.z - - dic_offset["mz"],
+                         0])
 
-    return np.append(vector, new_data), first_timestamp
+    return np.append(vector, [new_data], axis=0), first_timestamp
 
 
 def calc_data_mean(data):
@@ -52,39 +52,46 @@ def offset_calculation(dic):
     return dic_offset_mean
 
 
-def save_trainnning_data(data, categ, action_list):
+def save_trainnning_data(data):
 
-    path = "./../data/raw_learning_data"
+    path = "./../data/lstm_data"
 
     files = os.listdir(path)
 
-    vector_categ = np.ones((data.shape[0], 1)) * float(categ)
-
-    data = np.append(data, vector_categ, axis=1)
+    file_exists = False
 
     for file in files:
         if file.find("raw_learning_data.") != -1:
             prev_data_array = np.load(path + "/raw_learning_data.npy")
-            data = np.append(prev_data_array, data, axis=0)
+            data = np.append(prev_data_array, [data], axis=2)
+            file_exists = True
 
-    np.save(path + "/raw_learning_data.npy", data)
+    if file_exists is False:
+        data_aux = np.zeros((data.shape[0], data.shape[1], 0))
+        print(data_aux)
+        data_aux[:, :, 0] = data
+        np.save(path + "/raw_learning_data.npy", data_aux)
+    else:
+        np.save(path + "/raw_learning_data.npy", data)
 
-    idx_dic = {}
-    idx_list = []
-    for line in data:
-        key = str(int(line[-1]))
-        if not str(key) in idx_dic:
-            idx_dic[key] = 1
-            idx_list.append(key)
-        else:
-            idx_dic[key] += 1
 
-    idx_list.sort()
-
-    print(f"There are now {data.shape[0]} experiments in total")
-
-    for idx in idx_list:
-        print(f"category {action_list[int(idx)]}: {idx_dic[idx]} experiments")
+    #
+    # idx_dic = {}
+    # idx_list = []
+    # for line in data:
+    #     key = str(int(line[-1]))
+    #     if not str(key) in idx_dic:
+    #         idx_dic[key] = 1
+    #         idx_list.append(key)
+    #     else:
+    #         idx_dic[key] += 1
+    #
+    # idx_list.sort()
+    #
+    # print(f"There are now {data.shape[0]} experiments in total")
+    #
+    # for idx in idx_list:
+    #     print(f"category {action_list[int(idx)]}: {idx_dic[idx]} experiments")
 
 
 if __name__ == '__main__':
@@ -120,7 +127,7 @@ if __name__ == '__main__':
     rospy.init_node("training_data_aquisition", anonymous=True)
 
     data_for_learning = DataForLearning()
-    arm_gripper_comm = ArmGripperComm()
+
 
     rate = rospy.Rate(config["rate"])
 
@@ -130,23 +137,23 @@ if __name__ == '__main__':
     # -------------------------------INITIATE ROBOT------------------------------------------------
     # ---------------------------------------------------------------------------------------------
 
-    try:
-        joints = config["initial_pose"]
-        arm_gripper_comm.move_arm_to_joints_state(joints[0], joints[1], joints[2], joints[3], joints[4], joints[5])
-        if args["activate_gripper"] == 1:
-            input("Press ENTER to activate gripper in 3 secs")
-            for i in range(0, 3):
-                print(i + 1)
-                time.sleep(1)
-
-            arm_gripper_comm.gripper_init()
-            time.sleep(1.5)
-
-            arm_gripper_comm.gripper_close_fast()
-            time.sleep(0.5)
-            arm_gripper_comm.gripper_disconnect()
-    except:
-        print("ctrl+C pressed")
+    # try:
+    #     joints = config["initial_pose"]
+    #     arm_gripper_comm.move_arm_to_joints_state(joints[0], joints[1], joints[2], joints[3], joints[4], joints[5])
+    #     if args["activate_gripper"] == 1:
+    #         input("Press ENTER to activate gripper in 3 secs")
+    #         for i in range(0, 3):
+    #             print(i + 1)
+    #             time.sleep(1)
+    #
+    #         arm_gripper_comm.gripper_init()
+    #         time.sleep(1.5)
+    #
+    #         arm_gripper_comm.gripper_close_fast()
+    #         time.sleep(0.5)
+    #         arm_gripper_comm.gripper_disconnect()
+    # except:
+    #     print("ctrl+C pressed")
 
     list_gripper_calibration = []
     dic_offset_calibration = {"fx": [],
@@ -185,98 +192,100 @@ if __name__ == '__main__':
 
     limit = int(config["time"] * config["rate"])
 
-    trainning_data_array = np.empty((0, limit * len(config["data"])))
+    # trainning_data_array = np.empty((0, limit * len(config["data"])))
 
-    for j in range(0, args["number_reps"]):
+    # for j in range(0, args["number_reps"]):
+    #
+    #     print(f"Waiting for action to initiate experiment {j + 1}...")
 
-        print(f"Waiting for action to initiate experiment {j + 1}...")
+    print("Waiting action!!!!")
 
-        while not rospy.is_shutdown():
-            data_mean = calc_data_mean(data_for_learning)
-            variance = data_mean - rest_state_mean
-            # print(variance)
-            if abs(variance) > config["force_threshold_start"]:
-                break
-
-            time.sleep(0.1)
-
-        time.sleep(config["waiting_offset"]) # time waiting to initiate the experiment
-
-        # ---------------------------------------------------------------------------------------------
-        # -------------------------------------GET DATA------------------------------------------------
-        # ---------------------------------------------------------------------------------------------
-
-        end_experiment = False
-        first_time_stamp = None
-        vector_data = np.empty((0, 0))
-
-        i = 0
-        treshold_counter = 0
-
-        rate.sleep()  # The first time rate sleep was used it was giving problems (would not wait the right amout of time)
-
-        try:
-            while not rospy.is_shutdown() and i < limit:
-
-                i += 1
-                print(data_for_learning)
-                vector_data, first_time_stamp = add_to_vector(data_for_learning,
-                                                              vector_data, first_time_stamp, dic_variable_offset)
-
-                data_mean = calc_data_mean(data_for_learning)
-                variance = data_mean - rest_state_mean
-
-                # print(variance)
-                if abs(variance) < config["force_threshold_end"]:
-                    treshold_counter += 1
-                    if treshold_counter >= config["threshold_counter_limit"]:
-                        end_experiment = True
-                        break
-                else:
-                    treshold_counter = 0
-
-                rate.sleep()
-        except:
-            print("ctrl+C pressed")
-
-        if end_experiment:
-            print(f"Data collection interrupted in experiment {j + 1}")
-            trainning_data_array = trainning_data_array[:-1, :]
+    while not rospy.is_shutdown():
+        data_mean = calc_data_mean(data_for_learning)
+        variance = data_mean - rest_state_mean
+        # print(variance)
+        if abs(variance) > config["force_threshold_start"]:
             break
 
-        print(vector_data.shape)
-        # vector_norm = normalize_data(vector_data, limit)
-        trainning_data_array = np.append(trainning_data_array, [vector_data], axis=0)
+        time.sleep(0.1)
+
+    time.sleep(config["waiting_offset"]) # time waiting to initiate the experiment
+
+    # ---------------------------------------------------------------------------------------------
+    # -------------------------------------GET DATA------------------------------------------------
+    # ---------------------------------------------------------------------------------------------
+
+    end_experiment = False
+    first_time_stamp = None
+    vector_data = np.empty((0, 14))
+
+    i = 0
+    treshold_counter = 0
+
+    rate.sleep()  # The first time rate sleep was used it was giving problems (would not wait the right amout of time)
+
+    # try:
+    while not rospy.is_shutdown() and i < 600:
+
+        i += 1
+        # print(data_for_learning)
+        vector_data, first_time_stamp = add_to_vector(data_for_learning,
+                                                      vector_data, first_time_stamp, dic_variable_offset)
+
+        data_mean = calc_data_mean(data_for_learning)
+        variance = data_mean - rest_state_mean
+
+        # print(variance)
+        if abs(variance) < config["force_threshold_end"]:
+            treshold_counter += 1
+            if treshold_counter >= config["threshold_counter_limit"]:
+                end_experiment = True
+                break
+        else:
+            treshold_counter = 0
+
+        rate.sleep()
+    # except:
+    #     print("ctrl+C pressed")
+
+    # if end_experiment:
+    #     print(f"Data collection interrupted in experiment {j + 1}")
+    #     trainning_data_array = trainning_data_array[:-1, :]
+    #     break
+
+    print(vector_data.shape)
+    # vector_norm = normalize_data(vector_data, limit)
+    # trainning_data_array = np.append(trainning_data_array, [vector_data], axis=0)
 
 
-    if args["category"] == -1:
-        category = input("Input the number that corresponds to the category trainned: ")
-    else:
-        category = args["category"]
+    # if args["category"] == -1:
+    #     category = input("Input the number that corresponds to the category trainned: ")
+    # else:
+    #     category = args["category"]
 
-    pos = f'time: {datetime.now().strftime("%d/%m/%Y %H:%M:%S")}, action: {category}, ' \
-          f'number_experiments: {trainning_data_array.shape[0]}, ' \
+    pos = f'time: {datetime.now().strftime("%d/%m/%Y %H:%M:%S")}, ' \
+          f'number_experiments: {vector_data.shape[0]}, ' \
           f'joints_pos: {data_for_learning.joints_position}, gripper_pos: ' \
           f'({data_for_learning.wrench_pose.position.x}, ' \
           f'{data_for_learning.wrench_pose.position.y}, {data_for_learning.wrench_pose.position.z}, ' \
           f'{data_for_learning.wrench_pose.orientation.x}, {data_for_learning.wrench_pose.orientation.y}, ' \
           f'{data_for_learning.wrench_pose.orientation.z}, {data_for_learning.wrench_pose.orientation.w})'
 
-    del data_for_learning, arm_gripper_comm
+    del data_for_learning
 
     # ---------------------------------------------------------------------------------------------
     # -------------------------------SAVE TRAINNING DATA-------------------------------------------
     # ---------------------------------------------------------------------------------------------
 
-    classification = config["action_classes"][int(category)]
-    out = input(f"You wish to save the {trainning_data_array.shape[0]} {classification} experiment? (s/n)\n")
+    print(f"Trainning saved!\nIt was saved {vector_data.shape[0]} experiments")
+    save_trainnning_data(vector_data)
 
-    if out == "s":
-        with open('../data/raw_learning_data/position_historic.txt', 'a') as f:
-            f.write(pos + "\n")
-
-        print(f"Trainning saved!\nIt was saved {trainning_data_array.shape[0]} experiments")
-        save_trainnning_data(trainning_data_array, category, config["action_classes"])
-    else:
-        print("Trainning not saved!")
-
+    # if out == "s":
+    #     with open('../data/raw_learning_data/position_historic.txt', 'a') as f:
+    #         f.write(pos + "\n")
+    #
+    #     print(f"Trainning saved!\nIt was saved {vector_data.shape[0]} experiments")
+    #     save_trainnning_data(vector_data)
+    # else:
+    #     print("Trainning not saved!")
+    #
